Bug 1857287 - Move isProductUrl to ContentState
