+++ android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -827,6 +827,14 @@ sealed class ContentAction : BrowserAction() {
      * Indicates the given [tabId] was unable to be checked for form data.
      */
     data class CheckForFormDataExceptionAction(val tabId: String, val throwable: Throwable) : ContentAction()
+
+    /**
+     * Updates the [ContentState.isProductUrl] state for the non private tab with the given [tabId].
+     */
+    data class UpdateProductUrlStateAction(
+        val tabId: String,
+        val isProductUrl: Boolean,
+    ) : ContentAction()
 }
 
 /**
@@ -873,18 +881,6 @@ sealed class CookieBannerAction : BrowserAction() {
      */
     data class UpdateStatusAction(val tabId: String, val status: CookieBannerHandlingStatus) :
         CookieBannerAction()
-}
-
-/**
- * [BrowserAction] implementations related to updating the [SessionState.isProductUrl]
- * of a single [SessionState] inside [BrowserState]
- */
-sealed class ShoppingProductAction : BrowserAction() {
-    /**
-     * Updates the [SessionState.isProductUrl] state for the non private tab with the given [tabId].
-     */
-    data class UpdateProductUrlStatusAction(val tabId: String, val isProductUrl: Boolean) :
-        ShoppingProductAction()
 }
 
 /**
